Reorganize/Reorder storage related code
